Organize types and update analyzers to use constants and gitcommit
