+++ src/Moq.Analyzers/Common/CompilationExtensions.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.CodeAnalysis.Operations;
-
-namespace Moq.Analyzers;
+﻿namespace Moq.Analyzers.Common;
 
 internal static class CompilationExtensions
 {
